+++ NEWS.md
@@ -4,8 +4,9 @@
   * Feature: Handle more general vulnerabilities
   * Feature: Add `audit_description()` function
   * Feature: Add `expect_secure()` for the {testthat} package
+  * Feature: Use `~/.ossindex/.oss-index-config`, if it exists
   * Add Josiah Perry as contributor
-  * Minor: Link to [https://www.jumpingrivers.com](jumpingrivers.com) blog post
+  * Minor: Link to [jumpingrivers.com](https://www.jumpingrivers.com) blog post
   * Minor: Make `verbose = FALSE` completely silent
 
 # oysteR 0.0.3